add new implementation of LateBindingCalls based on revised logic
